fix: mocked TaskRow placeholder moved search inside, reordered utils, implemented my/team filtering
