+++ frontend/src/community/crm/constants/taskConstants.ts
@@ -2,23 +2,30 @@ import { CrmPriorityEnum } from "~community/crm/enums/common";
 
 export const TASK_SEARCH_DEBOUNCE_DELAY = 500;
 
-export const priorityOptions = [
+export const TASK_TAB_IDS = {
+  MY_TASKS: "my-tasks",
+  TEAM_TASKS: "team-tasks",
+  COMPLETED_TASKS: "completed-tasks"
+}
+
+export const SALES_REP_RESTRICTED_TASK_TABS = [
+  TASK_TAB_IDS.TEAM_TASKS
+];
+
+export const PRIORITY_OPTIONS = [
   {
-    label: "Low",
     key: "low",
     value: CrmPriorityEnum.LOW,
     backgroundColor: "bg-semantic-green-background",
     textColor: "text-semantic-green-text"
   },
   {
-    label: "Medium",
     key: "medium",
     value: CrmPriorityEnum.MEDIUM,
     backgroundColor: "bg-semantic-amber-background",
     textColor: "text-semantic-amber-text"
   },
   {
-    label: "High",
     key: "high",
     value: CrmPriorityEnum.HIGH,
     backgroundColor: "bg-semantic-red-background",
