+++ frontend/src/community/crm/hooks/useGetPriorityOptions.tsx
@@ -2,14 +2,14 @@ import { Label } from "@rootcodelabs/skapp-ui";
 import { useMemo } from "react";
 
 import { useTranslator } from "~community/common/hooks/useTranslator";
-import { priorityOptions } from "~community/crm/constants/taskConstants";
+import { PRIORITY_OPTIONS } from "~community/crm/constants/taskConstants";
 
 const useGetPriorityOptions = () => {
   const translateText = useTranslator("crmModule", "tasks", "addTaskModal");
 
   return useMemo(
     () =>
-      priorityOptions.map((option) => ({
+      PRIORITY_OPTIONS.map((option) => ({
         id: option.key,
         label: (
           <Label
