+++ frontend/src/community/crm/utils/taskUtils.ts
@@ -1,0 +1,26 @@
+import { CrmTaskDetailType } from "~community/crm/types/CommonTypes";
+
+import { isDueToday, isDueTomorrow, isOverdue } from "./taskValidations";
+
+export const groupTasksByDueDate = (tasks: CrmTaskDetailType[]) => {
+  const overdue: CrmTaskDetailType[] = [];
+  const dueToday: CrmTaskDetailType[] = [];
+  const dueTomorrow: CrmTaskDetailType[] = [];
+  const upcoming: CrmTaskDetailType[] = [];
+
+  for (const task of tasks) {
+    if (!task.dueAt) {
+      upcoming.push(task);
+    } else if (isOverdue(task.dueAt)) {
+      overdue.push(task);
+    } else if (isDueToday(task.dueAt)) {
+      dueToday.push(task);
+    } else if (isDueTomorrow(task.dueAt)) {
+      dueTomorrow.push(task);
+    } else {
+      upcoming.push(task);
+    }
+  }
+
+  return { overdue, dueToday, dueTomorrow, upcoming };
+};
